Android.mk: Add PN557 firmware
The PN557 firmware is used in the SoMC SM8150 Kumano platform
but there's a quirk: the NXP NFC HAL is requesting the firmware
as "libpn553_fw.so" instead of "libpn557_fw.so".

For this reason, declare the PN557 FW as libpn553_fw: this
anyway has a barrier for NXP_CHIP_FW_TYPE, so we cannot get
any conflict with the real pn553 anyway.

P.S.: The other way would be to declare it as LOCAL_MODULE
libpn557_fw, but then we would have to add an ifeq for the
NXP_CHIP_FW_TYPE also in the nxp-vendor.mk, where we select
the firmwares to add to the build.

Let's keep it simple!